requirements.txt
lxml==3.7.3
python_telegram_bot==5.3.0
requests==2.13.0
typing==3.6.1
peewee==2.10.1
PyMySQL==0.7.11
emoji==0.4.5
